updated notebook
